+++ src/stmem.c
@@ -213,7 +213,7 @@ stmemCopy(const mem_hdr * mem, off_t offset, char *buf, size_t size)
     char *ptr_to_buf = NULL;
     int bytes_from_this_packet = 0;
     int bytes_into_this_packet = 0;
-    debug(19, 6) ("memCopy: offset %d: size %d\n", offset, size);
+    debug(19, 6) ("memCopy: offset %d: size %d\n", (int) offset, size);
     if (p == NULL)
 	return 0;
     assert(size > 0);
